fix: address PR review feedback -- simplify auth helpers
- Move createConnectionFactory() from Constants to new ConnectionHelper class
- Simplify factory: drop isConfigured()/normalizeHost(), let SDK validate
- Reorder constants: Entra ID as Option 1 (recommended), connection string as Option 2
- Change defaults from placeholder strings to null
- Remove SharedAccessSignature validation (unsupported in 2.1.0)
- Update all 7 samples to use ConnectionHelper.createConnectionFactory()
